use pointers for user model
