+++ core/handlers.go
@@ -5,35 +5,11 @@ import (
 	"errors"
 	"net/http"
 	"strconv"
-	"strings"
 
-	"github.com/gorilla/context"
 	"github.com/julienschmidt/httprouter"
 
 	"github.com/iterableio/api/db"
 )
-
-func auth(h httprouter.Handle) httprouter.Handle {
-	return func(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
-		authHeader := r.Header.Get("Authorization")
-
-		// should have format `token {TOKEN}`
-		payload := strings.Split(authHeader, " ")
-		if len(payload) != 2 || payload[0] != "token" {
-			// do things
-			WriteErrorUnauthorized(w, errors.New("Missing/incorrect auth header"))
-			return
-		}
-		user, err := db.FindUserByToken(payload[1])
-		if err != nil {
-			WriteErrorUnauthorized(w, errors.New("User doesn't exist"))
-			return
-		}
-		context.Set(r, "user", user)
-		h(w, r, ps)
-		context.Clear(r)
-	}
-}
 
 func getUser(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
 	targetId, err := strconv.Atoi(ps.ByName("userId"))
@@ -42,7 +18,7 @@ func getUser(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
 		return
 	}
 
-	if currentUser := context.Get(r, "user").(db.User); currentUser.Id != targetId {
+	if currentUser := getCurrentUser(r); currentUser.Id != targetId {
 		WriteErrorUnauthorized(w, errors.New("You dont have permission to view this"))
 		return
 	}
